+++ packages/blockchain-wallet-v4-frontend/src/data/components/plugin/sagaRegister.ts
@@ -3,11 +3,12 @@ import { takeLatest } from 'redux-saga/effects'
 import sagas from './sagas'
 import { actions } from './slice'
 
-export default ({ api }) => {
-  const pluginSagas = sagas({ api })
+export default ({ api, coreSagas, networks }) => {
+  const pluginSagas = sagas({ api, coreSagas, networks })
 
   return function* pluginSaga() {
     yield takeLatest(actions.getPublicAddress, pluginSagas.getPublicAddress)
     yield takeLatest(actions.getWallet, pluginSagas.getWallet)
+    yield takeLatest(actions.autoLogin, pluginSagas.autoLogin)
   }
 }
