+++ packages/blockchain-wallet-v4-frontend/src/data/components/sagaRegister.ts
@@ -57,7 +57,7 @@ export default ({ api, coreSagas, networks }) =>
     yield fork(manageAddresses({ api, networks }))
     yield fork(onboarding())
     yield fork(nfts({ api, coreSagas, networks }))
-    yield fork(plugin({ api }))
+    yield fork(plugin({ api, coreSagas, networks }))
     yield fork(priceChart())
     yield fork(refresh())
     yield fork(request({ api, coreSagas, networks }))
