--- packages/blockchain-wallet-v4-frontend/src/data/pluginAuth/sagaRegister.ts
@@ -1,12 +1,0 @@
-import { takeLatest } from 'redux-saga/effects'
-
-import sagas from './sagas'
-import { actions } from './slice'
-
-export default ({ api, coreSagas, networks }) => {
-  const pluginLoginSaga = sagas({ api, coreSagas, networks })
-
-  return function* pluginAutoAuthSaga() {
-    yield takeLatest(actions.autoLogin, pluginLoginSaga.autoLogin)
-  }
-}
