--- packages/blockchain-wallet-v4-frontend/src/data/pluginAuth/slice.ts
@@ -1,16 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit'
-
-const initialState = {}
-
-const pluginAuthSlice = createSlice({
-  initialState,
-  name: 'pluginAuth',
-  reducers: {
-    autoLogin: () => {},
-    loginRoutineSaga: () => {}
-  }
-})
-
-const { actions } = pluginAuthSlice
-const pluginAuthReducer = pluginAuthSlice.reducer
-export { actions, pluginAuthReducer }
